+++ tests/test_names.py
@@ -11,4 +11,4 @@ def test_names(demo_data):
     xr_data: xr.Dataset = obj_mc._data
     df_export: pd.DataFrame = xr_data.mc.to_dataframe(original_column_names=True)
     for col in demo_data.columns:
-        assert col in list(df_export.columns), f'{col} is not in {list(demo_data.columns)}'
+        assert col in df_export.columns.to_list(), f'{col} is not in {demo_data.columns.to_list()}'